# 028. Swagger (API UI)/api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API for SHOP Website",
    "description": "This is a REST API application made with Express and documented with Swagger"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    },
    {
      "url": "http://shop.didan.id.vn/api"
    },
    {
      "url": "https://shop.didan.id.vn/api"
    }
  ],
...
  "paths": {
...
    "/admin/add-product": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint to create a new product",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Iphone 12"
                  },
                  "price": {
                    "type": "number",
                    "example": 1000
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "description": {
                    "type": "string",
                    "example": "A smartphone from Apple"
                  }
                },
                "required": [
                  "name",
                  "price",
                  "image",
                  "description"
                ]
              }
            }
          }
        }
      }
    },
...
  },
...
}